Fixes script-only mode on Windows (#18393)
Two main areas of change here - one to be nice to the user
and make sure that Ninja is used to build script-only projects,
as well as make sure its installed and help them install it.

The other area is to just skip over a bunch of checks that were
added to the Visual Studio stuff when script only mode is active.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>