Fixed Shader reload bug. (#18555)
* Fixed Shader reload bug.

There are three parts to this Pull Request:
1- Bug fix (Shader::OnAssetReloaded)
2- A new level called `ShaderReloadTest`, which contains assets + python
   script to validate the fix.
3- A new feature for O3DE that allows Gems and Game Projects to load custom
   PassTemplate assets without writing C++ code.

To understand the Bug Fix let's review the structure of ShaderAssets:
The `AZ::RPI::Shader`` class owns `AZ::RPI::ShaderAsset m_asset`.
This was the only asset that the Shader class was expecting to be notified
during Shader::OnAssetReloaded.
But, the ShaderAsset references several Root AZ::RPI::ShaderVariantAsset(s),
one for each Supervariant. The structure looks like:
```
AZ::RPI::ShaderAsset m_asset
  |
  |---> AZ::RPI::ShaderVariantAsset (Supervariant 0 Root variant)
  |---> AZ::RPI::ShaderVariantAsset (Supervariant 1 Root variant)
  | ...
  |---> AZ::RPI::ShaderVariantAsset (Supervariant N-1 Root variant)
```
When the user modified the Shader code, the AssetManager would dispatch
an `OnAssetReloaded` event for each asset in the chain shown above... But
the `Shader` class was only listening to `m_asset`. To make matters worst
the AssetManager wouldn't update the internal references to the Root
 ShaderVariantAssets.

The solution is two folds:
1. Register for OnAssetReloaded events for m_asset and for each of its Root
   ShaderVariantAsset(s).
2. When OnAssetReloaded() is called, keep track of all the new asset references
   in a dictionary. And only when the size of the Dictionary matches the total
   number of expected notifications then the Shader calls Init() again, and
   the dictionary is cleared and ready for the next time the user modifies the
   shader code.

In order to validate the bug fix, a new Level called `ShaderReloadTest` was
 added, which is a self contained set of assets and a Python script called
 `shader_reload_test.py`. To learn more about this script, please read the
 attached `README.md`. REMARK: This level is NOT a typical Python Automation
 Test, instead it is expected to be ran manually by engineers whenever they
 suspect that there are issues with Shader reloading.

Finally a new System Component called AZ::RPI::PassTemplatesAutoLoader was
added. It provides a very convenient opt-in service for Gems and Game Projects
to load custom PassTemplates. See `PassTemplatesAutoLoader.h` for details.
The O3DE wiki will be updated with description on how to use this service.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>