Added a small set of fixes that faciliate Level creation with Python.
1. Fixed bug in the MaterialBuilder where Material names
like "metal.001.material" were being produced as "metal.azmaterial".
With this fix the same source file is now produced as "metal.001.azmaterial".
This bug was messing with the posibility of finding a material product
asset from the path of the source asset.

2. Improved editor_python_test_tools/utils.py by Try/Excepting
the inclusion of the `multiplayer` module.
Not all projects use the Multiplayer Gem.

3. Added new Global Method to the BehaviorContext (for Automation only)
that allows to add the NonUniformScaleComponent to an entity.
`azlmbr.editor.AddNonUniformScaleComponent(entityId: EntityId, scale: Vector3)`

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>