Cherrypick from Dev - Fixes Slow Icon rendering
Chery-picks commit 40cdf65 from development.

* Fixes Slow Icon Rendering, Issue #18567

Fixes issue https://github.com/o3de/o3de/issues/18567

This batch renders icons.  Note that the old API was retained, for backward
compatibility, so all existing code in other gems will still work
(none were found in the engine), but the new API is preferred.

After this change, I can't really notice a difference between icons on and off
in terms of rendering for Startergame (A couple thousand icons).

Before this change, it would drop to about 4ps, from 60.

I'm sure there's always a way to improve this further but just this change
alone is enough for the icon rendering to no longer matter in terms of
profiler performance, so other things should probably be looked at before
we return to this.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>