Changed default memory allocator to basic OS malloc because HPHthe Editor crashes when using HPHA with Vulkan on levels with large amount of Meshes. The crash occured when loading levels or entering/exiting game mode.
Changed to TickBus::QueueFunction instead of SystemTickBus::QueueFunction
because it appears to not send deferred events when a different level is
loaded. After extensive validation it was found that with TickBus there
are no late events dispatched on behalf of destroyed components. With
SystemTickBus was easy to crash.

Added a python script:
Gems/EditorPythonBindings/Editor/Scripts/level_load_unload_stress.py
that helps automate and iterate across a list of levels.
Opens a level, enters/exits game mode, and can wait an arbitrary amount
of frames in between those operations.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>