Fixes race conditions and deadlocks when loading levels. (#18609)
* Fixes race conditions and deadlocks when loading levels.

Fixes #18597

The most important change in this PR is that InstanceDatabase::EmplaceInstance
has been changed so it manages when to lock `m_databaseMutex`. In particular,
m_databaseMutex is not locked anymore during `InstanceData<Type>` creation
because creating resources like StreamingImage(s) is a multi-threaded operation
coordinated by `AsyncUploadQueue` where secondary threads (like `Secondary Copy Queue`) sometimes release
other StreamingImage(s) and that operation also locks m_databaseMutex. This is a clear
deadlock case if m_databaseMutex was already locked on the Main Thread.

The second most important change pertains to some components that instantiate StreamingImage(s)
objects in the same thread when the AssetBus calls one of the OnAssetReady() or OnAssetReloaded()
functions. The AssetBus locks its own mutex. This can also cause deadlocks,
because when  AsyncUploadQueue() is called, and an StreamingImage
is released during the `Secondary Copy Queue` thread, the StreamingImage class would call:
```cpp
Data::AssetBus::MultiHandler::BusDisconnect(GetAssetId());
```
Which will deadlock when it tries to lock the AssetBus mutex.
The fix is to defer the OnAssetXXX functions on the TickBus queue.
REMARK:
This PR fixes:
1. ImageBasedLightComponentController.cpp OnAssetXXXX
2. MaterialComponentController.cpp OnAssetXXXX
3. TerrainSurfaceMaterialsListComponent.cpp OnAssetXXXX
BUT, any other component that instantiates StreamingImage(s) exactly during OnAssetXXX functions
would need to apply a similar fix to avoid deadlocks.

The 3rd kind of bugs that have been fixed is related with race condition crashes
when some `InstanceData<Type>` classes are created during multi-threaded scenarios
like MeshDrawPacket::DoUpdate(), which runs from multiple concurrent jobs. To avoid
the crashes a new mutex was added to `InstanceDatabase<Type>` called `m_newInstanceMutex`
which is only locked when a new `InstanceData<Type>` is being created and constructed.

Finally, a major improvement was done to `CryMT::queue` class by changing its underlying container
from `std::vector` to `std::queue`. I discovered the innefficiency when I added all kinds of verbose
AZ_Printf statements when debugging all the bugs mentioned above. And I noticed that the main reason
for slow debugging was because the old implementation using `std::vector` was calling:
```cpp
v.erase(v.begin());
```
Which, for a vector, is a big no-no because erasing from the begginning of a vector causes long copy operations.
Hard to believe we've been dealing with this lingering performance threat for so long.

Added a python script:
Gems/EditorPythonBindings/Editor/Scripts/level_load_unload_stress.py
that helps automate and iterate across a list of levels.
Opens a level, enters/exits game mode, and can wait an arbitrary amount
of frames in between those operations.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>