Fixed PreviewRenderer material update crash. (#18643)
Fixes #18629
Fixes #18630

The crash was originated by `AtomToolsFramework::PreviewRenderer`. And
the easiest way to reproduce was by making changes to an active material
asset using the MaterialEditor. Order of events for the crash:
The MaterialComponentController owned by AtomToolsFramework::PreviewRenderer,
detects a change in the material asset and it enqueues
MaterialComponentController::InitializeMaterialInstancePostTick in the TickBus
which captures the `this` address. Then immediately `MaterialComponentController::Deactivate()` is called.
Later, in the next TickBus event, `MaterialComponentController::InitializeMaterialInstancePostTick`
is executed with the same `this` pointer that was destroyed in the previous frame.

Solution, added a std:queue<Asset> to MaterialComponentController where
all notified assets are stored. The lambda function was removed so
there's no storage of the `this` pointer anymore. Then, when the
MaterialComponentController is Deactivated, all queued assets are safely
released and there's no lambda function to execute in the next Tick.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>