Fixes many issues discovered by the Address Sanitizer
With these changes, I was able to run all of the tests in the following folders
in debug, with Address Sanitizer enabled.
* Everything in Code/* - so framework, tools, AP, etc.
* Everything In Gems/Atom* - so all atom tests in there.

Note that you will not be able to reproduce this - the current version of GoogleMock
has ASAN issues inside it.  I will be following this up with a fix to GoogleMock
that upgrades it to the latest version, which has those ASAN issues fixed.

However, doing so requires mass renaming some macros, and thus the changes that
would cause would dirty up this PR and make it nearly impossible to properly review.

Once this is merged, I will follow up with the GoogleMock upgrade.

If you have any questions about any individual change, please submit a comment
to the PR.  Each of these changes were related to an actual ASAN error or other
test failure and were discussed in the Discord as they were found.

please don't ask me to split this up.  Running tests under ASAN is kind of an all
or nothing thing - ASAN kill the program the moment the first one is encountered,
making it incredibly time consuming and error prone to try to split these up, and
many of these core problems are inter-related.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>