MaterialComponentRequestBus::IsMaterialAssetReady
Added two new methods to MaterialComponentRequestBus:
1. IsMaterialAssetReady(const MaterialAssignmentId& materialAssignmentId);
2. IsDefaultMaterialAssetReady(const MaterialAssignmentId& materialAssignmentId);

This is useful because at runtime you may not get the actual
value of a particular material property if the MaterialAsset associated
to a particular MaterialAssignmentId has not been loaded in memory.

For a Lua or Python script, once IsMaterialAssetReady() returns `true`.
it is guaranteed that any of the GetPropertyValueXXX() methods will actually return
a non-default value.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>