CherryPick from Dev:  ASAN fixes (from PR #18655)
* Fixes many issues discovered by the Address Sanitizer

With these changes, I was able to run all of the tests in the following folders
in debug, with Address Sanitizer enabled.
* Everything in Code/* - so framework, tools, AP, etc.
* Everything In Gems/Atom* - so all atom tests in there.

Note that you will not be able to reproduce this - the current version of GoogleMock
has ASAN issues inside it.  I will be following this up with a fix to GoogleMock
that upgrades it to the latest version, which has those ASAN issues fixed.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

* Also fixes an issue where if the build machine is in a locale not using period as the
decimal separator, it will fail the tests.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>