Improves the performance of making trivial edits in a large level
Improves
https://github.com/o3de/o3de/issues/18656

It turns out that when you make a trivial edit, so for example, modifying
a simple entity property like a property field or a transform, the system
already updates the instance being edited (its cached json DOM) and also
already updates the template it belongs to (the prefab file in memory that
the instance came from), and also, the actual entity being modified.

There is no other cached data to update that I'm aware of, except if there
are OTHER instances of the same prefab in the scene.  It adds those
instances to a queue to re-propogate the changes to them.

The main thing that this does is skip the already-upated instance from
being added to that queue.  It is already up to date.

The side effect of this is that when you have a large level and you
are modifying individual entities in them, what would previously happen
is that it would add all instances to the update queue, including the
one that was just updated.  And in the case of a large level, the one
that was just updated *is the level itself* leading to a respawn of
the entire level.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>