Fixes a bad cherry-pick merge that would prevent ASAN working
The cherry-pick merge of ASAN fixes from dev to 24092.2
caused an ifdef to shift around.

Note that the ifdef shifted around in exactly a way that
meant that not having ASAN enabled (which is the default)
worked just fine and behaved normally and didn't cause
any actual changes to the code as compiled or executed.

But if you turn on ASAN, it won't compile since it switches
to malloc allocator and this ifdef eats most of the file.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>