Improved ShaderStreamBufferViewsInterface by adding the ability to query for the `IndexBufferView` and `StreamBufferView` which will help the developer find the actual Offsets and ByteCounts for each BufferView which can be use in shader constants to find the correct data in the ByteAddressBuffers.
Added convenient API to BufferView to get
the BindlessSrg indices for READ and RW (optional).

Added opt-in `DrawSrg::m_modelLodMeshIndex` shader
constant. To opt-in the developer must enable
the shader macro `USE_DRAWSRG_MESHLOD_MESHINDEX`
and MeshDrawPacket will set its value to the
subMesh index. By default it is disabled so it doesn't
affect material pipelines. This is useful for
Texture Space Shading techniques where it can
be important to know which subMesh index a DrawItem
is referring to.

Extended EditorMeshComponent.cpp to list
the number of Triangles and Vertices per subMesh
in the readonly Mesh statistics.

Improved MaterialTypeBuilder.cpp by removing the
";" extra character per line in "objectSrgAddtion".
This enables materialpipeline files to add
custom structs to ObjectSrg. This also implies
that it is up to the developer to manually add ";"
where needed.

Add more macro protections in *.azsl* files
to prevent usage of functions only available to
 fragment shaders like ddx_fine or clip() or discard
 with help of `#ifdef CS_SAMPLERS` which means
 the shader being compiled is a Compute shader.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>