Fixes a issue #18720 (vulkan malloc crash) and compile error (#18736)
* Fixes a issue #18720 and a compile error with latest Windows SDK

Fixes:
https://github.com/o3de/o3de/issues/18720

I was unable to test this without also fixing a compiler error
introduced by the latest windows SDK.

The latest windows SDK changed the basic definition of the macro
`INFINITY` from being a huge number on its own, to be two huge
numbers multiplied by each other.  This causes the compiler to error
out with a 'math overflow' whenever INFINITY is used anywhere.

One of our tests uses it as an input number to test whether it can
detect an infinite matrix.  I've bracketed the test with a
push and pop for that warning.  It only happens on MSVC and only with
latest SDK.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

* Fixes typo

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

---------

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>