BUGFIX - Crash on CTRL+G if SC is open.
Script canvas would crash (debug) or perform
undefined behavior when you pressed CTRL+G and the
script canvas window was open.

This was because the CMake file applied a preprocessor
definition to targets in ScriptCanvas as PRIVATE, and that
preprocessor definition changed the size of a type
(It removed a member variable).

The problem is, that header also contains a templated
function from a static library that will be compiled in other
compile units, not just the private target in ScriptCanvas,
causing undefined behavior.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>