Possible fix for mouse cursor and ui background missing
My analysis showed that for non-indexed rendering, in both
vulkan and dx12, the code that was supposed to set the vertex
buffer was not being called.  The CommandList class that is
responsible for setting them doesn't do anything if there are
no index buffers.

This is a tenative fix which uses the other buffers (ie, the
vertex buffer) if its told to set the vertex buffer with no
index buffers present.

I was able to repro the problem in both vulkan and dx12.
This fixes the problem in both vulkan and dx12.

Fixes https://github.com/o3de/o3de/issues/18769
Fixes https://github.com/o3de/o3de/issues/18890

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>