Moves ImGuiPass into the public API section
Atom Feature Common was recently split into public and static
private.  This particular interface used to be public
and is now private.

This change is the bare minimum I could do in order to
make existing code function with the same API as before.

I don't think the underlying structure is good...
I would prefer if for example the ImGUI stuff was
all in the ImGUI Gem instead, but this fixes the
regression at least.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>