Improves AP startup time and CPU Usage
* MaterialBuilder no longer specifies that all textures are critical
* AssetProcessor escalates jobs that are blocking critical jobs from starting
* AssetProcessor (GUI mode only) reserves half jof the CPU cores for
  critical jobs.  When no critical jobs are running, it uses fewer cores
  to give more CPU usage to the editor to maintain framerates.
* User settings like "Skip startup scan" are now project-specific instead
  of global.
* New user setting: Verbose logging (off by default), reduces log spam
  by about 90% when enabled.

  Timing WITH new changes (starting editor with no cache)
  *   42s (  42s total) analyzing
  * 2m45s (3m28s total) "Asset Processor working..."
  * 1m15s (4m44s total) 60fps interactive editor with defaultlevel (meshes still compiling)

Timing WITHOUT the new changes (starting editor with no cache)
  *   42s (  42s total) analyzing (I did not change analyze so this makes sense)
  * 3m34s (4m18s total) "Asset processor working..."
  * 1m22s (5m40s total) 10-20 fps interactive editor with defaultlevel (meshes still compiling)

So about a 45s saving on "Asset Processor Working"
and about 56s total saving with a 60fps editor while it works in background.

Both timings were captured with all other variables hte same, and were
performed on a "hot" file cache, as in, run it once, then clear cache
and immediately run it again to get official timings.    The timing
on a cold cache for the new changes is even more of a difference than
a cold cache without the changes.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>