Removes deprecated ConsoleHelpGen system (#19041)
The console help gen system came along wiht CryEngine's
legacy codebase, and has never worked properly in O3DE.

The Console has a command to dump out all commands and their
parameters.

When you execute this command, it writes a txt file with all
the commands and their parameters, and on windows, also
attempts to write out a CHM file.  The CHM file generation
has always failed and/or crashed.  CHM is a deprecated format
for the last decade or so.

https://en.wikipedia.org/wiki/Microsoft_Compiled_HTML_Help

Note that this just removes the CHM generation, not the txt
generation, that still works fine.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>