Fixes the installer o3de.sh script working on linux. (#19045)
When the installer (.deb) installs on linux, it
creates a symlink to the o3de.sh script in the user's
bin directory (on the path) so the user can just run `o3de.sh`
from anywhere.

However, the script needs to know where the actual o3de
python path is, relative to the script, and thus needs to
de-symlink like all the other scripts (like python.sh and such).

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>