Fixes a crash that occurs when meshes update in Raytrace (#19084)
The raytrace feature processor was incorrectly using
`auto` instead of `auto &` when iterating over its
structures.  Most of the time, this would only be a slight
performance hit, but fatally, when a mesh was cleared
out of the raytracing feature processor, it used
```cpp
for (auto [deviceIndex, uncompactedBlasEnqueuedForDeletion] : m_uncompactedBlasEnqueuedForDeletion)
{
    uncompactedBlasEnqueuedForDeletion.erase(id);
}
```

The lack of the ampersand meand that the erase was being carried out
on a copy of the map, which would not actually remove the
item from the original map, leaving bogus entries in there.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>