Fixes material jobs depending on the wrong platform
When a job is emitted by a builder, it can depend on other jobs.
If it depends on other jobs, it waits for those jobs to finish before it
starts its own work.

When your job specifies a dependency, it specifies a tuple of
(platform, job key, source file).  The job key is just the type of job,
a constant string like "Material Builder".

When asset processor tells a builder to emit jobs, it supplies
it a list of platforms that are currently enabled for the project,
for example "pc", "android", "ios", etc.

The Material builder and Material Type builder were creating ONE
job descriptor and then emitting it for each platform as a copy.

The builder was trying to update the platform field in the dependency
but unfortunately it only checked if the platform field was blank.

So in a loop where we go through eahc platform, it would set
the job being emitted to the platform of the first job, and then
on each subsequent platform, it would not update the platform field
because it was not blank anymore, and thus would be emitting jobs
for one platform, but with dependencies for a different platform.

This caused those other jobs to get started before the actual files
they depend on were actually ready (since they would only wait) for
the original platform's job to finish.

This change fixes that by always setting the platform field
to the current platform, so that the dependencies are always
correctly set to the platform of the job being emitted.

It also adds a check to the Asset Processor Manager to reject and
auto fail jobs that have dependencies that are not for the current platform
to prevent this from happening in the future.

It also adds several new automated tests to make sure that the
triggering of dependent jobs works correctly, with, and without the
subids being specified.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>