Fixes `main` tests running a periodic test which fails
Two fixes here.  One is that the `periodic` test
was running when it should not, becuase it was not
marked with SUITE_periodic.

The other is that the test itself had an invalid
asset reference to the bunny model.  Its no clear
to me why the reference changed, but it did.  This is
a very ancient asset, so its possible that it changed
some time in the deep past.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>