Updates miniaudio to latest
1. Adds a function to 3rdparty.cmake to handle 3rd party fetchcontent
   libraries, following DRY principle
2. Updates miniaudio to latest stable release.
3. Ensures that the vorbis decoder plugin is initialized and present

Note that miniaudio used to be a header-only library and used stb's
"Single header file" vorbis to provide ogg/vorbis support.

The latest miniaudio is a real full static library and header, and uses
the real vorbis and ogg.  While miniaudio doesn't force you to invoke
cmake `install` commands, vorbis and ogg do, so we had to set the
EXCLUDE_FROM_ALL property for them, which prevents their `install`
commands from executing.  This also prevents them from being added
to the dependency tree at all, so `add_dependencies` had to be invoked.

Other than that, this is a pretty standard 3rd party fetchcontent
situation, and should serve as a model for how to do others like this.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>