When patching a 3rd party library, this (#19184)
adds a check to see if it can find `git` and if it
cannot, tells the user that it cannot, instead
of failing mysteriously.

(This was encountered by a user who had git installed
on MSYS command line, but not in their actual windows
PATH).

Example of the error that appears now:
```
--   RecastNavigation Gem uses https://github.com/recastnavigation/recastnavigation.git commit 5a870d4 (License: Zlib)
  MSBuild version 17.14.14+a129329f1 for .NET Framework

    Performing update step for 'recastnavigation-populate'
    -- Already at requested ref: 5a870d427e47abd4a8e4ce58a95582ec049434d5
    Performing patch step for 'recastnavigation-populate'
    -- Could NOT find Git (missing: GIT_EXECUTABLE)
    CMake Error at C:/o3de-worktrees/stab/cmake/PatchIfNotAlreadyPatched.cmake:34 (message):
      Git executable not found.  Please make sure `git` is in your PATH

```

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>