Fixes crashes and ASAN errors in AzCore benchmarks
I was trying to benchmark the streamer stuff to compare
it for the new PR that improves it, and found these
benchmarks to be horribly broken and full of memory problems.

AFter these fixes, you can run the benchmarks in debug
or profile, ASAN or not ASAN, and it will run to the end.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>