Use setup-gradle to ensure that the version of gradle we use is always the one we specify.
Gradle 9.0.0 has issues with O3DE - the NDK must be newer
than the one we use in order to work with Gradle 9.

However, the newer NDK has other compile issues with O3DE.

Until we fix them, we can set the version of gradle to this
specific version using the setup-gradle action.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>