Replaces name fields of assets in Particle with Asset References
The product file data (as in, the compiled asset) was always using
asset references (`AZ::Data::Asset<Material> etc`) but the source json
files were using string path names.

This made it much more expensive at compile time and validation time,
for example, if you assign "Sphere.azmodel" it was previously trying
to make sure that "sphere.fbx" was compiled,by replacing the extension
"azmodel" with "fbx".  However, its possible that "sphere.azmodel"
came from "sphere.obj", or "sphere.usd", or any other model format.

It's also possible that "sphere.azmodel" came from a model that
was not named "sphere" at all, for example, if you had a model
named "shapes.fbx", that contained a cube, and a sphere.

This change makes it so that the source files also use Asset References,
which means that the Asset Processor can directly validate that the
asset exists, and that it is of the correct type, and does not need to
do any guessing based on file names.

It also replaces the custom asset control with the standard AZ Framework
asset picker and control, as well as uses the tools busses to convert
from the old "name" format to the new "asset reference" format, so all
old assets still function as before.

In order to convert them, I needed to make 2 changes to the core,
the first is there was simply a bug where the API to look up Asset
products by the source file name was not actually filling out the
entire response structure, leaving the file name empty.

The second is a convenience function to do a best guess conversion of
a file name to an asset reference, by doing a series of lookups from
the fastest most reliable to the slowest "Guess".

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>