Fix monolithic builds trying to build dummy O3DE Launcher app
The O3DE.GameLauncher target is a sort of universal
empty application script-only-mode uses to launch the game.

Because script-only mode cannot use a compiler or linker,
it cannot use monolithic mode.  It makes no sense to generate
these targets in monolithic mode, as they would be missing
all the gems that they depend on.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>