NickCodeReviewPass_1
Updated Entity and GameEntityContext comments.
Made the ActiveState methods in Entity more overt. For brainstorming purposes. Just for proper clarity on what each is for.
Renamed variables to m_ and s_ for static.
Removed the "Evaluate" bool in the setters, now you must call ApplyEffectiveState after making changes. This pushes the responsibility to the handlers over the Entity itself.
- Up to debate.
Updated EntityActiveSystemComponent to match the m_ and s_ variable naming.
Updated the GameEntityContextComponent:
- Removed a redundant method.
- Removed all of the added Utility methods. They were useless. AZStd::erase, FindEntity event, and a direct unordered_set[] operator call fixed it all.
- Updated all Entity State Handling to ApplyEffectiveState, as that is nolonger being done by the Entity.
Tried my best to preserve proper formatting, and line endings. Here's hoping.

Left the Transform/Parent methods and behaviour the same, pending a serious discussion on Transform functionality.

Signed-off-by: GaianHelmers <gaianhelmers@genomestudios.ca>